use component host in graph aspect instead of workspace / scope directly move ComponentGraph.buildFromLegacy (static function) to the graph builder
