+++ scopes/component/graph/graph-builder.ts
@@ -1,50 +1,45 @@
 import { ComponentID } from '@teambit/component';
-import { ScopeMain } from '@teambit/scope';
-import { Workspace } from '@teambit/workspace';
-import {
-  buildOneGraphForComponents,
-  buildOneGraphForComponentsUsingScope,
-} from 'bit-bin/dist/scope/graph/components-graph';
+import { ComponentFactory } from '@teambit/component';
 import { ComponentGraph } from './component-graph';
+import { Graph as LegacyGraph } from 'bit-bin/dist/scope/graph/graph';
+import { Dependency } from './model/dependency';
 
 export class GraphBuilder {
   _graph?: ComponentGraph;
   _initialized = false;
-  constructor(private workspace?: Workspace, private scope?: ScopeMain) {}
+  constructor(private componentHost: ComponentFactory) {}
 
   async getGraph(ids?: ComponentID[]): Promise<ComponentGraph | undefined> {
-    if (this.workspace) {
-      if (!ids) ids = (await this.workspace.list()).map((comp) => comp.id);
+    const legacyGraph = await this.componentHost.getLegacyGraph(ids);
+    const graph = await this.buildFromLegacy(legacyGraph);
 
-      const bitIds = ids.map((id) => id._legacy);
-      const legacyGraph = await buildOneGraphForComponents(bitIds, this.workspace.consumer);
-      const graph = await ComponentGraph.buildFromLegacy(legacyGraph, this.workspace);
-
-      this._graph = graph;
-      this._initialized = true;
-      return this._graph;
-    }
-
-    // Build graph from scope
-    if (this.scope) {
-      if (!ids) ids = (await this.scope.list()).map((comp) => comp.id);
-
-      const bitIds = ids.map((id) => {
-        let bitId = id._legacy;
-        // The resolve bitId in scope will remove the scope name in case it's the same as the scope
-        // We restore it back to use it correctly in the legacy code.
-        if (!bitId.hasScope()) {
-          bitId = bitId.changeScope(this.scope?.name);
-        }
-        return bitId;
-      });
-      const legacyGraph = await buildOneGraphForComponentsUsingScope(bitIds, this.scope.legacyScope);
-      const graph = await ComponentGraph.buildFromLegacy(legacyGraph, this.scope);
-
-      this._graph = graph;
-      this._initialized = true;
-      return this._graph;
-    }
+    this._graph = graph;
+    this._initialized = true;
     return this._graph;
+  }
+
+  private async buildFromLegacy(legacyGraph: LegacyGraph): Promise<ComponentGraph> {
+    const newGraph = new ComponentGraph();
+
+    const setNodeP = legacyGraph.nodes().map(async (nodeId) => {
+      const componentId = await this.componentHost.resolveComponentId(nodeId);
+      const component = await this.componentHost.get(componentId);
+      if (component) {
+        newGraph.setNode(componentId.toString(), component);
+      }
+    });
+    await Promise.all(setNodeP);
+
+    const setEdgePromise = legacyGraph.edges().map(async (edgeId) => {
+      const source = await this.componentHost.resolveComponentId(edgeId.v);
+      const target = await this.componentHost.resolveComponentId(edgeId.w);
+      const edgeObj =
+        legacyGraph.edge(edgeId.v, edgeId.w) === 'dependencies' ? new Dependency('runtime') : new Dependency('dev');
+      newGraph.setEdge(source.toString(), target.toString(), edgeObj);
+    });
+    await Promise.all(setEdgePromise);
+
+    newGraph.versionMap = newGraph._calculateVersionMap();
+    return newGraph;
   }
 }
