+++ scopes/component/graph/graph.provider.ts
@@ -1,16 +1,14 @@
 import { GraphqlMain } from '@teambit/graphql';
-import { ScopeMain } from '@teambit/scope';
-import { Workspace } from '@teambit/workspace';
 import { ComponentMain } from '@teambit/component';
 
 import { GraphBuilder } from './graph-builder';
 import { graphSchema } from './graph.graphql';
 
-export type GraphDeps = [GraphqlMain, ComponentMain, Workspace, ScopeMain];
+export type GraphDeps = [GraphqlMain, ComponentMain];
 
-export async function provide([graphql, componentMain, workspace, scope]: GraphDeps) {
-  const graphBuilder = new GraphBuilder(workspace, scope);
+export async function provide([graphql, componentMain]: GraphDeps) {
   const host = componentMain.getHost();
+  const graphBuilder = new GraphBuilder(host);
   graphql.register(graphSchema(graphBuilder, host));
   return graphBuilder;
 }
